requirements.txt
# Locked requirements
openai==1.3.7
aiohttp==3.9.3
cchardet==2.1.7
gTTS==2.5.1
openai==1.12.0
playsound==1.3.0
PyAudio==0.2.14
pydub==0.25.1
pynput==1.7.6
pyperclip==1.8.2
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
rumps==0.4.0
simpleaudio==1.0.4
six==1.16.0
tiktoken==0.6.0
tqdm==4.66.2
langchain==0.1.6
